+++ Eventy/src/main/java/org/example/eventy/solutions/models/Service.java
@@ -1,10 +1,9 @@
 package org.example.eventy.solutions.models;
 
-import jakarta.annotation.Nullable;
 import org.example.eventy.common.models.ReservationConfirmationType;
 
 public class Service extends Solution {
-    private ServiceHistory currentService;
+    private ServiceHistory previousService;
 
     private String specifics;
     private int minReservationTime;
@@ -26,16 +25,16 @@ public class Service extends Solution {
         this.reservationType = reservationType;
     }
 
-    public Service(ServiceHistory currentService) {
-        this.currentService = currentService;
+    public Service(ServiceHistory previousService) {
+        this.previousService = previousService;
     }
 
-    public ServiceHistory getCurrentService() {
-        return currentService;
+    public ServiceHistory getPreviousService() {
+        return previousService;
     }
 
-    public void setCurrentService(ServiceHistory currentService) {
-        this.currentService = currentService;
+    public void setPreviousService(ServiceHistory currentService) {
+        this.previousService = currentService;
     }
 
     public String getSpecifics() {
